+++ src/pybamm/solvers/idaklu_solver.py
@@ -76,6 +76,8 @@ class IDAKLUSolver(pybamm.BaseSolver):
                 "increment_factor": 1.0,
                 # Enable or disable linear solution scaling
                 "linear_solution_scaling": True,
+                # Silence Sundials errors during solve
+                "silence_sundials_errors": False,
                 ## Main solver
                 # Maximum order of the linear multistep method
                 "max_order_bdf": 5,
@@ -176,6 +178,7 @@ class IDAKLUSolver(pybamm.BaseSolver):
             "epsilon_linear_tolerance": 0.05,
             "increment_factor": 1.0,
             "linear_solution_scaling": True,
+            "silence_sundials_errors": False,
             "max_order_bdf": 5,
             "max_num_steps": 100000,
             "dt_init": 0.0,
@@ -681,13 +684,17 @@ class IDAKLUSolver(pybamm.BaseSolver):
         atol = self._check_atol_type(atol, y0full.size)
 
         timer = pybamm.Timer()
-        solns = self._setup["solver"].solve(
-            t_eval,
-            t_interp,
-            y0full,
-            ydot0full,
-            inputs,
-        )
+        try:
+            solns = self._setup["solver"].solve(
+                t_eval,
+                t_interp,
+                y0full,
+                ydot0full,
+                inputs,
+            )
+        except ValueError as e:
+            # Return from None to replace the C++ runtime error
+            raise pybamm.SolverError(str(e)) from None
         integration_time = timer.time()
 
         return [
@@ -734,14 +741,15 @@ class IDAKLUSolver(pybamm.BaseSolver):
             termination = "final time"
         elif sol.flag < 0:
             termination = "failure"
+            msg = idaklu.sundials_error_message(sol.flag)
             match self._on_failure:
                 case "warn":
                     warnings.warn(
-                        f"FAILURE {self._solver_flag(sol.flag)}, returning a partial solution.",
+                        msg + ", returning a partial solution.",
                         stacklevel=2,
                     )
                 case "raise":
-                    raise pybamm.SolverError(f"FAILURE {self._solver_flag(sol.flag)}")
+                    raise pybamm.SolverError(msg)
 
         if sol.yp.size > 0:
             yp = sol.yp.reshape((number_of_timesteps, number_of_states)).T
@@ -787,7 +795,7 @@ class IDAKLUSolver(pybamm.BaseSolver):
                 time_indep = True
 
             newsol._variables[var] = pybamm.ProcessedVariableComputed(
-                [model.variables_and_events[var]],
+                [model.get_processed_variable_or_event(var)],
                 base_variables,
                 [data],
                 newsol,
@@ -1002,40 +1010,3 @@ class IDAKLUSolver(pybamm.BaseSolver):
             t_interp=t_interp,
         )
         return obj
-
-    @staticmethod
-    def _solver_flag(flag):
-        flags = {
-            99: "IDA_WARNING: IDASolve succeeded but an unusual situation occurred.",
-            2: "IDA_ROOT_RETURN: IDASolve succeeded and found one or more roots.",
-            1: "IDA_TSTOP_RETURN: IDASolve succeeded by reaching the specified stopping point.",
-            0: "IDA_SUCCESS: Successful function return.",
-            -1: "IDA_TOO_MUCH_WORK: The solver took mxstep internal steps but could not reach tout.",
-            -2: "IDA_TOO_MUCH_ACC: The solver could not satisfy the accuracy demanded by the user for some internal step.",
-            -3: "IDA_ERR_FAIL: Error test failures occurred too many times during one internal time step or minimum step size was reached.",
-            -4: "IDA_CONV_FAIL: Convergence test failures occurred too many times during one internal time step or minimum step size was reached.",
-            -5: "IDA_LINIT_FAIL: The linear solver's initialization function failed.",
-            -6: "IDA_LSETUP_FAIL: The linear solver's setup function failed in an unrecoverable manner.",
-            -7: "IDA_LSOLVE_FAIL: The linear solver's solve function failed in an unrecoverable manner.",
-            -8: "IDA_RES_FAIL: The user-provided residual function failed in an unrecoverable manner.",
-            -9: "IDA_REP_RES_FAIL: The user-provided residual function repeatedly returned a recoverable error flag, but the solver was unable to recover.",
-            -10: "IDA_RTFUNC_FAIL: The rootfinding function failed in an unrecoverable manner.",
-            -11: "IDA_CONSTR_FAIL: The inequality constraints were violated and the solver was unable to recover.",
-            -12: "IDA_FIRST_RES_FAIL: The user-provided residual function failed recoverably on the first call.",
-            -13: "IDA_LINESEARCH_FAIL: The line search failed.",
-            -14: "IDA_NO_RECOVERY: The residual function, linear solver setup function, or linear solver solve function had a recoverable failure, but IDACalcIC could not recover.",
-            -15: "IDA_NLS_INIT_FAIL: The nonlinear solver's init routine failed.",
-            -16: "IDA_NLS_SETUP_FAIL: The nonlinear solver's setup routine failed.",
-            -20: "IDA_MEM_NULL: The ida mem argument was NULL.",
-            -21: "IDA_MEM_FAIL: A memory allocation failed.",
-            -22: "IDA_ILL_INPUT: One of the function inputs is illegal.",
-            -23: "IDA_NO_MALLOC: The ida memory was not allocated by a call to IDAInit.",
-            -24: "IDA_BAD_EWT: Zero value of some error weight component.",
-            -25: "IDA_BAD_K: The k-th derivative is not available.",
-            -26: "IDA_BAD_T: The time t is outside the last step taken.",
-            -27: "IDA_BAD_DKY: The vector argument where derivative should be stored is NULL.",
-        }
-
-        flag_unknown = "Unknown IDA flag."
-
-        return flags.get(flag, flag_unknown)